requirements.txt
b3j0f.conf>=0.3.13
six>=1.10.0
parse>=1.6.6
link.json>=0.4
link.wsgi>=0.2
